Skip transload if no matching file fields
If the exclude selector results in an empty set of file field elements,
don't bother opening the transloadit dialog. This is especially helpful
if you pass an exclude argument like:

exclude: function() { return !$(this).val(); }

a form with an empty upload field will just pass through and not open a
transloadit dialog box.